fix(state): 식사 완료 횟수의 정수 범위 확장

+++ include/philo.h
@@ -25,7 +25,7 @@ typedef struct s_config
 typedef struct s_philo
 {
 	int				id;
-	int				meals;
+	int64_t			meals;
 	int64_t		last_meal_ms;
 	pthread_t		thread;
 	pthread_mutex_t	*left_fork;
